mcp-medical-server/requirements.txt
mcp>=0.9.0
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
authlib>=1.3.0
bcrypt>=4.1.0
pyjwt>=2.8.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0

# ML/AI Models for document processing (offline-capable)
openai-whisper>=20230314
face-recognition>=1.3.0
dlib>=19.24
pytesseract>=0.3.10
Pillow>=10.0.0
easyocr>=1.7.0

# AI Brain - LLM Integration
openai>=1.0.0
anthropic>=0.18.0
langdetect>=1.0.9
slowapi>=0.1.9
